lxd/instance: add riscv64 QEMU definitions
Add missing definitions to run virtual RISC-V machines.
This is enough to build an LXD snap that supports RISC-V VMs.

For RISC-V there is no trusted party to sign shim for all operating systems
like Microsoft does for x86 and ARM. As LXD defaults to secure boot add
fake entries for the firmware.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>