lxd/instance: add riscv64 QEMU definitions
Add missing definitions to run virtual RISC-V machines.
This is enough to build an LXD snap that supports RISC-V VMs.

Secure boot support for riscv64 does not exist, yet.
security.secureboot=false needs to be set in the profile or configuration.

ACPI support as of QEMU 8.2.2 is not sufficient to boot Ubuntu images.
acpi=off enables booting via device-trees.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>