driver_qemu: informative error message for QEMU version
When reading the QEMU version fails a numeric error code is the only clue
to the reason why a failure occurred. It is preferable to provide the
output to stderr.

In an instance that I experienced the new message was

    Failed getting version during QEMU initialization:
    /snap/lxd/x2/bin/qemu-system-riscv64:
    error while loading shared libraries: libatomic.so.1:
    cannot open shared object file: No such file or directory

instead of

    Failed getting version during QEMU initialization: 127

With the new message it becomes clear why qemu-system-riscv64 cannot be
executed.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>